+++ app/pb_kits/playbook/pb_section_separator/_section_separator.jsx
@@ -1,21 +1,27 @@
 /* @flow */
 import React from 'react'
 import classnames from 'classnames'
-
 import Caption from '../pb_caption/_caption.jsx'
 
 type SectionSeparatorProps = {
   className: String,
   text: String,
+  orientation?: 'horizontal' | 'vertical',
   variant?: 'card' | 'background',
+  dark?: Boolean
 }
 
 const SectionSeparator = ({
   className,
   text,
+  orientation = 'horizontal',
   variant = 'card',
+  dark = false,
 }: SectionSeparatorProps) => {
-  const css = classnames([`pb_section_separator_kit_${variant}`, className])
+  const themeStyle = dark === true ? '_dark' : ''
+  const css = classnames([
+    `pb_section_separator_kit_${variant}_${orientation}` + themeStyle,
+    className])
 
   return (
     <div className={css}>