+++ spec/pb_kits/playbook/kits/weekday_stacked_spec.rb
@@ -13,7 +13,7 @@ RSpec.describe Playbook::PbWeekdayStacked::WeekdayStacked do
                       .with_default(false) }
   it { is_expected.to define_prop(:date)
                       .with_default(Date.current) }
-  it { is_expected.to define_prop(:day_only)
+  it { is_expected.to define_prop(:variant)
                       .with_default(false) }
 
   describe "#classname" do
@@ -30,7 +30,7 @@ RSpec.describe Playbook::PbWeekdayStacked::WeekdayStacked do
       expect(subject.new({date: Date.parse("2019/11/4")}).day_of_week).to eq '<time datetime="2019-11-04T00:00:00-05:00">Mon</time>'
       expect(subject.new({compact: true, date: Date.parse("2019/11/4")}).day_of_week).to eq '<time datetime="2019-11-04T00:00:00-05:00">M</time>'
       expect(subject.new({date: Date.parse("2019/11/4")}).formatted_month_and_day).to eq '<time datetime="2019-11-04T00:00:00-05:00">11/4</time>'
-      expect(subject.new({date: Date.parse("2019/11/4"), day_only: true}).formatted_month_and_day).to eq '<time datetime="2019-11-04T00:00:00-05:00">4</time>'
+      expect(subject.new({date: Date.parse("2019/11/4"), variant: "expanded"}).formatted_month_and_day).to eq '<time datetime="2019-11-04T00:00:00-05:00">Nov 4</time>'
     end
   end
 end