+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # Unreleased
 - New Kit - File Upload Kit ([#730](https://github.com/powerhome/playbook/pull/730) @thestephenmarshall)
   - Make doc tabs intelligently sense whether an example is available for either Rails or React versions of a kit and show accordingly
+- Added vertical section separator variants & dark version ([#729](https://github.com/powerhome/playbook/pull/729)@christinaatai)
 
 ## [4.11.0] 2020-4-10
 - Added Filter only & Sort only templates to filter kit + React Version ([#702](https://github.com/powerhome/playbook/pull/702)@kre8sions)